--- src/service.rs
@@ -1,1 +1,0 @@
-pub mod job_creation_service;
